web/flask-server/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.0
tbb==2021.13.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.57
ultralytics-thop==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3

# PyTorch with CUDA 12.1
torch==2.3.1+cu121

# Use the extra index URL to find the packages
--extra-index-url https://download.pytorch.org/whl/cu121
